# Build shape expressions and remove outputs that are only used by `aten::size`s