--- a/torch/csrc/jit/passes/tensorexpr_fuser.cpp
+++ b/torch/csrc/jit/passes/tensorexpr_fuser.cpp
@@ -29,17 +29,17 @@ bool isSupportedForBlock(Node* node) {
   }
 }
 
-bool usedOnlyInSize(Value *v) {
-  const auto &uses = v->uses();
-  return std::all_of(uses.begin(), uses.end(), [](const Use &u) {
+bool usedOnlyInSize(Value* v) {
+  const auto& uses = v->uses();
+  return std::all_of(uses.begin(), uses.end(), [](const Use& u) {
     return u.user->matches("aten::size(Tensor self) -> int[]");
   });
 }
 
-Value *broadcastSizes(at::ArrayRef<Value *> sizes, AliasDb *db) {
+Value* broadcastSizes(at::ArrayRef<Value*> sizes, AliasDb* db) {
   AT_ASSERT(!sizes.empty());
-  Graph *graph = sizes[0]->owningGraph();
-  Node *broadcast_n =
+  Graph* graph = sizes[0]->owningGraph();
+  Node* broadcast_n =
       graph->insertNode(graph->create(prim::BroadcastSizes, sizes));
   broadcast_n->output()->setType(ListType::ofInts());
   db->createValue(broadcast_n->output());
@@ -307,14 +307,12 @@ class TensorExprFuser {
   // Builds up expressions that compute shapes of all intermediates (and
   // outputs) of the fusion group, based on the sizes of inputs. You should run
   // DCE to remove those that you end up not using.
-  std::unordered_map<Value *, Value *>
-  buildShapeExpressions(Node *fusion_group) {
-
+  std::unordered_map<Value*, Value*> buildShapeExpressions(Node* fusion_group) {
     GRAPH_DUMP("buildShapeExpressions for ", fusion_group->g(attr::Subgraph));
     WithInsertPoint insert_guard{fusion_group->next()};
-    std::unordered_map<Value *, Value *> shape_of;
+    std::unordered_map<Value*, Value*> shape_of;
 
-    Graph *graph = fusion_group->owningGraph();
+    Graph* graph = fusion_group->owningGraph();
     auto subgraph = fusion_group->g(attr::Subgraph);
 
     auto inputs = fusion_group->inputs();
@@ -322,10 +320,13 @@ class TensorExprFuser {
     AT_ASSERT(inputs.size() == sinputs.size());
     for (size_t i = 0; i < inputs.size(); ++i) {
       if (inputs[i]->type()->isSubtypeOf(TensorType::get())) {
-        Value *soutput = graph->insert(aten::size, {inputs[i]});
+        Value* soutput = graph->insert(aten::size, {inputs[i]});
         aliasDb_->createValue(soutput);
-        GRAPH_DEBUG("Adding a mapping for %", sinputs[i]->debugName(), " ",
-                    getHeader(soutput->node()));
+        GRAPH_DEBUG(
+            "Adding a mapping for %",
+            sinputs[i]->debugName(),
+            " ",
+            getHeader(soutput->node()));
         shape_of[sinputs[i]] = soutput;
       }
     }
@@ -340,12 +341,12 @@ class TensorExprFuser {
     for (size_t i = 0; i < outputs.size(); ++i) {
       if (usedOnlyInSize(outputs[i]))
         continue;
-      Value *soutput = graph->insert(aten::size, {outputs[i]});
+      Value* soutput = graph->insert(aten::size, {outputs[i]});
       aliasDb_->createValue(soutput);
       shape_of[soutputs[i]] = soutput;
     }
 
-    for (Node *n : subgraph->nodes()) {
+    for (Node* n : subgraph->nodes()) {
       // XXX: Use of shape_of.emplace is crucial to the output shape
       // optimization!
       if (n->kind() == aten::cat) {
@@ -359,49 +360,50 @@ class TensorExprFuser {
         continue;
       }
       if (n->kind() == prim::ConstantChunk) {
-        Node *sizes_node = graph->insertNode(
+        Node* sizes_node = graph->insertNode(
             graph->create(prim::ChunkSizes, shape_of.at(n->input()), 2));
         sizes_node->i_(attr::dim, n->i(attr::dim));
         sizes_node->i_(attr::chunks, n->i(attr::chunks));
-        for (Value *output : sizes_node->outputs()) {
+        for (Value* output : sizes_node->outputs()) {
           aliasDb_->createValue(output);
         }
-        Value *regular_size = sizes_node->outputs().at(0);
-        Value *last_size = sizes_node->outputs().at(1);
+        Value* regular_size = sizes_node->outputs().at(0);
+        Value* last_size = sizes_node->outputs().at(1);
         regular_size->setType(ListType::ofInts());
         last_size->setType(ListType::ofInts());
         auto outputs = n->outputs();
-        for (Value *o : outputs.slice(0, outputs.size() - 1)) {
+        for (Value* o : outputs.slice(0, outputs.size() - 1)) {
           shape_of.emplace(o, regular_size);
         }
         shape_of.emplace(outputs.at(outputs.size() - 1), last_size);
         continue;
       }
-      auto tensor_inputs = filter(n->inputs(), [](Value *v) {
+      auto tensor_inputs = filter(n->inputs(), [](Value* v) {
         return v->type()->isSubtypeOf(TensorType::get());
       });
       GRAPH_DEBUG("Building sizes for ", getHeader(n));
       bool all_inputs_have_sizes = true;
-      auto shapes =
-          fmap(tensor_inputs, [&](Value *v) { 
-            GRAPH_DEBUG("Getting aten::size for %", v->debugName());
-            all_inputs_have_sizes &= shape_of.count(v);
-            return shape_of[v]; 
-            });
-      
+      auto shapes = fmap(tensor_inputs, [&](Value* v) {
+        GRAPH_DEBUG("Getting aten::size for %", v->debugName());
+        all_inputs_have_sizes &= shape_of.count(v);
+        return shape_of[v];
+      });
+
       if (!all_inputs_have_sizes) {
-        GRAPH_DEBUG("Not all tensor arguments have sizes available to compute the broadcasted size", getHeader(n));
+        GRAPH_DEBUG(
+            "Not all tensor arguments have sizes available to compute the broadcasted size",
+            getHeader(n));
         continue;
       }
-      shape_of.emplace(n->output(),
-                       shapes.size() == 1
-                           ? shapes[0]
-                           : broadcastSizes(shapes, aliasDb_.get()));
+      shape_of.emplace(
+          n->output(),
+          shapes.size() == 1 ? shapes[0]
+                             : broadcastSizes(shapes, aliasDb_.get()));
     }
     return shape_of;
   }
 
-  void removeOutputsUsedOnlyInSize(Node *fusion_group) {
+  void removeOutputsUsedOnlyInSize(Node* fusion_group) {
     if (fusion_group->kind() != prim::TensorExprGroup)
       return;
     auto subgraph = fusion_group->g(attr::Subgraph);
@@ -913,7 +915,7 @@ class TensorExprFuser {
     }
   }
 
-  void guardFusionGroupsAndRemoveOutputs(Block *block) {
+  void guardFusionGroupsAndRemoveOutputs(Block* block) {
     std::vector<Node*> fusion_groups;
     for (Node* n : block->nodes()) {
       for (Block* b : n->blocks()) {