--- a/torch/csrc/jit/passes/tensorexpr_fuser.h
+++ b/torch/csrc/jit/passes/tensorexpr_fuser.h
@@ -29,6 +29,9 @@ TORCH_API void RemoveProfileNodesAndSpecializeTypes(
     std::shared_ptr<Graph>& graph);
 TORCH_API void RemoveTensorTypeSpecializations(std::shared_ptr<Graph>& graph);
 
+TORCH_API bool usedOnlyInSize(Value *v);
+TORCH_API Value *broadcastSizes(at::ArrayRef<Value *> sizes, AliasDb *db);
+
 namespace tensorexpr {
 TORCH_API bool isSupported(Node* node);
 }