@bertmaher 2020-09-23
For posterity, here's the perf improvement of this diff on the fastrnns:
| name                                             | base time (s) |   diff time (s) |   % change |
| :---                                             |          ---: |            ---: |       ---: |
| lstm[aten]:bwd                                   |     29.191425 |       29.228951 |       0.1% |
| lstm[aten]:fwd                                   |     25.109489 |       25.232128 |       0.5% |
| lstm[cudnn]:bwd                                  |     17.512569 |       17.594995 |       0.5% |
| lstm[cudnn]:fwd                                  |     11.208224 |       11.301911 |       0.8% |
| lstm[jit]:bwd                                    |     26.906675 |       25.277245 |      -6.1% |
| lstm[jit]:fwd                                    |     15.593824 |       14.901732 |      -4.4% |
| lstm[jit_multilayer]:bwd                         |     25.605623 |       29.035940 |      13.4% |
| lstm[jit_multilayer]:fwd                         |     15.557957 |       15.140351 |      -2.7% |
| lstm[jit_premul]:bwd                             |     26.539135 |       23.700302 |     -10.7% |
| lstm[jit_premul]:fwd                             |     16.277058 |       13.827213 |     -15.1% |
| lstm[jit_premul_bias]:bwd                        |     24.006948 |       23.389751 |      -2.6% |
| lstm[jit_premul_bias]:fwd                        |     15.454223 |       13.576311 |     -12.2% |
| lstm[jit_simple]:bwd                             |     24.800737 |       24.927588 |       0.5% |
| lstm[jit_simple]:fwd                             |     15.306452 |       14.831862 |      -3.1% |
| lstm[py]:bwd                                     |     48.450756 |       52.208534 |       7.8% |
| lstm[py]:fwd                                     |     37.644066 |       39.758385 |       5.6% |